Fix the case the directory containing the domain authentication properties doesn't exist (com/stratelia/silverpeas/authentication for example). Fix the JMS topic names so that they match with their JNDI name

+++ src/main/scripts/groovy/updateAuthProperties.groovy
@@ -6,18 +6,21 @@ final String AuthPropertiesPath    = "${SILVERPEAS_HOME}/properties/org/silverpe
 def updateAllAuthPropertiesIn(String directory) {
   final String TempSuffix            = '.tmp'
   final String TemplateToReplace     = 'com.stratelia.'
-  new File(directory).eachFile { properties ->
-    if (properties.isFile()) {
-      FileWriter updatedProperties = new FileWriter(properties.path + TempSuffix)
-      new FileReader(properties).transformLine(updatedProperties) { line ->
-        if (line.contains(TemplateToReplace))
-          line = line.replaceAll(TemplateToReplace, 'org.')
-        if (!line.contains('SQLPasswordEncryption'))
-          line
+  File authDir = new File(directory)
+  if (authDir.exists() && authDir.isDirectory()) {
+    authDir.eachFile { properties ->
+      if (properties.isFile()) {
+        FileWriter updatedProperties = new FileWriter(properties.path + TempSuffix)
+        new FileReader(properties).transformLine(updatedProperties) { line ->
+          if (line.contains(TemplateToReplace))
+            line = line.replaceAll(TemplateToReplace, 'org.')
+          if (!line.contains('SQLPasswordEncryption'))
+            line
+        }
+        properties.delete()
+        new File(properties.path + TempSuffix).renameTo(properties)
       }
-      properties.delete()
-      new File(properties.path + TempSuffix).renameTo(properties)
-    }
+    } 
   } 
 }
